* origin/released:
  xserver libX11 libxcb mesa xcb-proto glproto git update March 31 2017

+++ mesalib/src/util/bitscan.c
@@ -32,7 +32,7 @@
 #elif defined(_MSC_VER) && (_M_IX86 || _M_ARM || _M_AMD64 || _M_IA64)
 #else
 int
-ffs(unsigned i)
+ffs(int i)
 {
    int bit = 0;
    if (!i)
@@ -63,7 +63,7 @@ ffs(unsigned i)
 #elif defined(_MSC_VER) && (_M_AMD64 || _M_ARM || _M_IA64)
 #else
 int
-ffsll(uint64_t val)
+ffsll(long long int val)
 {
    int bit;
